requirements.txt
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
beautifulsoup4==4.12.3
cachetools==5.3.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
cloudpickle==3.0.0
crc32c==2.3.post0
decorator==4.4.2
docker-pycreds==0.4.0
dt-apriltags==3.1.7
einops==0.7.0
evdev==1.7.0
filelock==3.9.0
fsspec==2023.12.2
furniture-bench==0.1
gitdb==4.0.11
GitPython==3.1.41
google-api-core==2.16.1
google-auth==2.27.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
gym==0.21.0
gym-notices==0.0.8
huggingface-hub==0.20.3
hydra-core==1.3.2
idna==3.4
imageio==2.33.1
imageio-ffmpeg==0.4.9
Jinja2==3.1.2
joblib==1.3.2
llvmlite
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
networkx
numba
numpy
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
proglog==0.1.10
protobuf==3.20.3
psutil==5.9.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pynput==1.7.6
pyrealsense2==2.54.2.5684
python-xlib==0.33
PyYAML==6.0.1
regex==2023.12.25
requests==2.28.1
rich==13.7.0
rsa==4.9
safetensors==0.4.2
sentry-sdk==1.40.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
soupsieve==2.5
sympy==1.12
tokenizers==0.15.1
torch==2.2.0+cu118
torchaudio==2.2.0+cu118
torchdata==0.7.1
torchvision==0.17.0+cu118
tqdm==4.66.1
transformers==4.37.2
triton==2.2.0
typing_extensions==4.8.0
urllib3==1.26.13
wandb==0.16.2
